Add more table to the database. Create backservice to update the user stock value

+++ Stock Manager Simulator Backend/Stock Manager Simulator Backend/Dtos/StockPriceDto.cs
@@ -12,6 +12,6 @@ namespace Stock_Manager_Simulator_Backend.Dtos
         public float DayHigh { get; set; }
         public float DayLow { get; set; }
         public float DayOpen { get; set; }
-        public int UpdateTimeInTimestamp { get; set; }
+        public long UpdateTimeInTimestamp { get; set; }
     }
 }
